Python: Split `def` and `use` nodes
This demonstrates a way of splitting up the external API of API graphs
so that definitions and uses are distinguished. The internal
representation is unchanged.

With this change, we can enforce the use of `getParameter`/`getArgument`
and other methods to conform to whether the given node is a definition
or a use.